Add reconciliation logic for Group
For now, it is expected that Group spec will contain
a list of strings containing the cluster names.
Based upon that list, the controller creates secondary PlacementRule
CRs for each cluster.
Down the line the rules engine tied to the Group will be implemented
and the PlacementRule objects created will honour that.